rename system to init
