+++ src/init/init.rs
@@ -1,0 +1,5 @@
+use log::*;
+use std::ffi::CString;
+use syslog::{BasicLogger, Facility, Formatter3164};
+
+use super::*;
