+++ src/lib.rs
@@ -40,8 +40,11 @@ use std::path::Path;
 use std::path::PathBuf;
 
 use anyhow::Context;
+use init::init_pid1_logging;
+use init::init_rootfs;
+use init::init_syslog_logging;
+use init::print_logo;
 use log::*;
-use system::{init::*, *};
 use tokio::net::UnixListener;
 use tokio_stream::wrappers::UnixListenerStream;
 use tonic::transport::{Certificate, Identity, Server, ServerTlsConfig};
@@ -52,10 +55,10 @@ use crate::observe::ObserveService;
 // use crate::runtime::LocalRuntimeService;
 
 mod codes;
+mod init;
 mod meta;
 mod observe;
 mod runtime;
-mod system;
 
 pub const AURAE_SOCK: &str = "/var/run/aurae/aurae.sock";
 
@@ -157,11 +160,11 @@ impl SystemRuntime {
     }
 
     fn init_pid_gt_1(&self) {
-        system::init::init_syslog_logging(self.logger_level);
+        init_syslog_logging(self.logger_level);
     }
 
     pub fn init(&self) {
-        if system::get_pid() == 1 {
+        if init::get_pid() == 1 {
             self.init_pid1();
         } else {
             self.init_pid_gt_1();
